+++ projects/amdsmi/tests/amd_smi_test/main.cc
@@ -24,7 +24,6 @@
 #include "amd_smi/impl/amd_smi_utils.h"
 #include "functional/api_support_read.h"
 #include "functional/computepartition_read_write.h"
-#include "functional/cross_process_serialization.h"
 #include "functional/err_cnt_read.h"
 #include "functional/evt_notif_read_write.h"
 #include "functional/fan_read.h"
@@ -42,7 +41,6 @@
 #include "functional/memory_read_write.h"
 #include "functional/memorypartition_read_write.h"
 #include "functional/metrics_counter_read.h"
-#include "functional/mutual_exclusion.h"
 #include "functional/overdrive_read.h"
 #include "functional/overdrive_read_write.h"
 #include "functional/pci_read_write.h"
@@ -253,6 +251,7 @@ TEST(amdsmitstReadOnly, TestAPISupportRead) {
   RunGenericTest(&tst);
 }
 
+/*
 TEST(amdsmitstReadOnly, TestMutualExclusion) {
   TestMutualExclusion tst;
   SetFlags(&tst);
@@ -262,15 +261,7 @@ TEST(amdsmitstReadOnly, TestMutualExclusion) {
   tst.Run();
   RunCustomTestEpilog(&tst);
 }
-
-TEST(amdsmitstReadOnly, TestCrossProcessSerialization) {
-  TestCrossProcessSerialization tst;
-  SetFlags(&tst);
-  tst.DisplayTestInfo();
-  tst.SetUp();
-  tst.Run();
-  RunCustomTestEpilog(&tst);
-}
+*/
 
 TEST(amdsmitstReadWrite, TestComputePartitionReadWrite) {
   if (!amd::smi::is_sudo_user()) GTEST_SKIP_("Invalid permission - Must run as super user");