+++ projects/rocshmem/src/gda/ibv_wrapper.hpp
@@ -77,6 +77,11 @@ class IBVWrapper {
     int modify_qp(struct ibv_qp *qp, struct ibv_qp_attr *attr, int attr_mask);
     int destroy_qp(struct ibv_qp *qp);
 
+    uint16_t flow_label_to_udp_sport(uint32_t fl);
+
+    struct ibv_ah* create_ah(struct ibv_pd *pd, struct ibv_ah_attr *attr);
+    int destroy_ah(struct ibv_ah *ah);
+
   private:
     struct ibv_funcs_t {
       struct ibv_device** (*get_device_list)(int *num_devices);
@@ -115,6 +120,9 @@ class IBVWrapper {
       struct ibv_qp* (*create_qp)(struct ibv_pd *pd, struct ibv_qp_init_attr *qp_init_attr);
       int (*modify_qp)(struct ibv_qp *qp, struct ibv_qp_attr *attr, int attr_mask);
       int (*destroy_qp)(struct ibv_qp *qp);
+
+      struct ibv_ah* (*create_ah)(struct ibv_pd *pd, struct ibv_ah_attr *attr);
+      int (*destroy_ah)(struct ibv_ah *ah);
     };
 
     /**